+++ app/dashboard/src/layouts/AssetsTable.tsx
@@ -336,7 +336,7 @@ export default function AssetsTable(props: AssetsTableProps) {
   const setAssetPanelProps = useSetAssetPanelProps()
 
   const hiddenColumns = columnUtils
-    .getColumnList(user, backend.type)
+    .getColumnList(user, backend.type, category)
     .filter((column) => !enabledColumns.has(column))
   const [sortInfo, setSortInfo] =
     React.useState<sorting.SortInfo<columnUtils.SortableColumn> | null>(null)
@@ -359,9 +359,6 @@ export default function AssetsTable(props: AssetsTableProps) {
   })
 
   const organization = organizationQuery.data
-
-  const isAssetContextMenuVisible =
-    category.type !== 'cloud' || user.plan == null || user.plan === backendModule.Plan.solo
 
   const nameOfProjectToImmediatelyOpenRef = React.useRef(initialProjectName)
   const [localRootDirectory] = localStorageProvider.useLocalStorageState('localRootDirectory')
@@ -802,6 +799,8 @@ export default function AssetsTable(props: AssetsTableProps) {
   const nodeMapRef = React.useRef<
     ReadonlyMap<backendModule.AssetId, assetTreeNode.AnyAssetTreeNode>
   >(new Map<backendModule.AssetId, assetTreeNode.AnyAssetTreeNode>())
+  const isAssetContextMenuVisible =
+    category.type !== 'cloud' || user.plan == null || user.plan === backendModule.Plan.solo
 
   const queryClient = useQueryClient()
 
@@ -1850,8 +1849,20 @@ export default function AssetsTable(props: AssetsTableProps) {
               nonConflictingProjectCount={projects.length - conflictingProjects.length}
               doUpdateConflicting={(resolvedConflicts) => {
                 for (const conflict of resolvedConflicts) {
-                  fileMap.set(conflict.current.id, conflict.file)
-                  void doUploadFile(conflict.current, 'update')
+                  const isUpdating = conflict.current.title === conflict.new.title
+
+                  const asset = isUpdating ? conflict.current : conflict.new
+
+                  fileMap.set(
+                    asset.id,
+                    new File([conflict.file], asset.title, {
+                      type: conflict.file.type,
+                      lastModified: conflict.file.lastModified,
+                    }),
+                  )
+
+                  insertAssets([asset], event.parentId)
+                  void doUploadFile(asset, isUpdating ? 'update' : 'new')
                 }
               }}
               doUploadNonConflicting={() => {
@@ -2254,6 +2265,29 @@ export default function AssetsTable(props: AssetsTableProps) {
     ],
   )
 
+  React.useEffect(() => {
+    // In some browsers, at least in Chrome 126,
+    // in some situations, when an element has a
+    // 'container-size' style, and the parent element is hidden,
+    // the browser can't calculate the element's size
+    // and thus the element doesn't appear when we unhide the parent.
+    // The only way to fix that is to force browser to recalculate styles
+    // So the trick is to change a property, trigger style recalc(`getBoundlingClientRect()`)
+    // and remove the property.
+    // since everything is happening synchronously, user won't see a broken layout during recalculation
+    if (!hidden && rootRef.current) {
+      for (let i = 0; i < rootRef.current.children.length; i++) {
+        const element = rootRef.current.children[i]
+
+        if (element instanceof HTMLElement) {
+          element.style.width = '0px'
+          element.getBoundingClientRect()
+          element.style.width = ''
+        }
+      }
+    }
+  }, [hidden])
+
   // This is required to prevent the table body from overlapping the table header, because
   // the table header is transparent.
   const updateClipPath = useOnScroll(() => {
@@ -2640,8 +2674,10 @@ export default function AssetsTable(props: AssetsTableProps) {
 
   const columns = React.useMemo(
     () =>
-      columnUtils.getColumnList(user, backend.type).filter((column) => enabledColumns.has(column)),
-    [backend.type, enabledColumns, user],
+      columnUtils
+        .getColumnList(user, backend.type, category)
+        .filter((column) => enabledColumns.has(column)),
+    [backend.type, category, enabledColumns, user],
   )
 
   const headerRow = (