+++ app/gui2/src/components/shared/AgGridTableView.vue
@@ -13,20 +13,18 @@ import { useAutoBlur } from '@/util/autoBlur'
 import type {
   CellEditingStartedEvent,
   CellEditingStoppedEvent,
+  ColDef,
+  ColGroupDef,
   ColumnResizedEvent,
   FirstDataRenderedEvent,
+  GetRowIdFunc,
+  GridApi,
   GridReadyEvent,
   RowDataUpdatedEvent,
   RowEditingStartedEvent,
   RowEditingStoppedEvent,
-  SortChangedEvent,
-} from 'ag-grid-community'
-import type {
-  ColDef,
-  ColGroupDef,
-  GetRowIdFunc,
-  GridApi,
   RowHeightParams,
+  SortChangedEvent,
 } from 'ag-grid-enterprise'
 import { type ComponentInstance, reactive, ref, shallowRef, watch } from 'vue'
 
@@ -57,7 +55,7 @@ const gridApi = shallowRef<GridApi<TData>>()
 const popupParent = document.body
 useAutoBlur(() => grid.value?.$el)
 
-function onGridReady(event: GridReadyEvent) {
+function onGridReady(event: GridReadyEvent<TData>) {
   gridApi.value = event.api
 }
 
@@ -92,15 +90,15 @@ watch(
 )
 
 function updateColumnWidths(event: FirstDataRenderedEvent | RowDataUpdatedEvent) {
-  if (event.columnApi == null) {
-    console.warn('AG Grid column API does not exist.')
+  if (event.api == null) {
+    console.warn('AG Grid API does not exist.')
     return
   }
-  const cols = event.columnApi.getAllGridColumns().filter((c) => {
+  const cols = event.api.getAllGridColumns().filter((c) => {
     const id = c.getColId()
     return id && !widths.has(id)
   })
-  event.columnApi.autoSizeColumns(cols)
+  event.api.autoSizeColumns(cols)
 }
 
 function lockColumnSize(e: ColumnResizedEvent) {