+++ app/gui2/src/components/visualizations/ScatterplotVisualization.vue
@@ -53,6 +53,8 @@ interface Data {
   focus: Focus | undefined
   points: PointsConfiguration
   data: Point[]
+  isTimeSeries: boolean
+  x_value_type: string
   is_multi_series?: boolean
 }
 
@@ -63,7 +65,7 @@ interface Focus {
 }
 
 interface Point {
-  x: number
+  x: number | DateObj | Date
   y: number
   label?: string
   color?: string
@@ -79,6 +81,7 @@ interface PointsConfiguration {
 enum ScaleType {
   Linear = 'linear',
   Logarithmic = 'logarithmic',
+  Time = 'time',
 }
 
 interface AxisConfiguration {
@@ -95,6 +98,15 @@ interface Color {
   red: number
   green: number
   blue: number
+}
+
+interface DateObj {
+  day?: number
+  month?: number
+  year?: number
+  hour?: number
+  minute?: number
+  second?: number
 }
 </script>
 
@@ -128,35 +140,65 @@ const SHAPE_TO_SYMBOL: Record<string, d3.SymbolType> = {
   triangle: d3.symbolTriangle,
 }
 
-const SCALE_TO_D3_SCALE: Record<ScaleType, () => d3.ScaleContinuousNumeric<number, number>> = {
-  [ScaleType.Linear]: () => d3.scaleLinear(),
-  [ScaleType.Logarithmic]: () => d3.scaleLog(),
+const createDateTime = (x: DateObj) => {
+  const dateTime = new Date()
+  if (x.day != null) dateTime.setDate(x.day)
+  if (x.month != null) dateTime.setMonth(x.month)
+  if (x.year != null) dateTime.setFullYear(x.year)
+  if (x.hour != null) dateTime.setHours(x.hour)
+  if (x.minute != null) dateTime.setMinutes(x.minute)
+  if (x.second != null) dateTime.setSeconds(x.second)
+  return dateTime
 }
 
 const data = computed<Data>(() => {
   let rawData = props.data
   const unfilteredData =
     Array.isArray(rawData) ? rawData.map((y, index) => ({ x: index, y })) : rawData.data ?? []
-  const data: Point[] = unfilteredData.filter(
-    (point) =>
-      typeof point.x === 'number' &&
-      !Number.isNaN(point.x) &&
-      typeof point.y === 'number' &&
-      !Number.isNaN(point.y),
-  )
+  let data: Point[]
+  // eslint-disable-next-line camelcase
+  const isTimeSeries: boolean =
+    'x_value_type' in rawData ?
+      rawData.x_value_type === 'Time' ||
+      rawData.x_value_type === 'Date' ||
+      rawData.x_value_type === 'Date_Time'
+    : false
+  // eslint-disable-next-line camelcase
+  if (isTimeSeries) {
+    data = unfilteredData
+      .filter((point) => typeof point.y === 'number' && !Number.isNaN(point.y))
+      .map((point) => ({ ...point, x: createDateTime(point.x as DateObj) }))
+  } else {
+    data = unfilteredData.filter(
+      (point) =>
+        typeof point.x === 'number' &&
+        !Number.isNaN(point.x) &&
+        typeof point.y === 'number' &&
+        !Number.isNaN(point.y),
+    )
+  }
   if (Array.isArray(rawData)) {
     rawData = {}
   }
   const axis: AxesConfiguration = rawData.axis ?? {
-    x: { label: '', scale: ScaleType.Linear },
+    x: { label: '', scale: isTimeSeries ? ScaleType.Time : ScaleType.Linear },
     y: { label: '', scale: ScaleType.Linear },
   }
   const points = rawData.points ?? { labels: 'visible' }
   const focus: Focus | undefined = rawData.focus
   // eslint-disable-next-line camelcase
   const is_multi_series: boolean = !!rawData.is_multi_series
-  // eslint-disable-next-line camelcase
-  return { axis, points, data, focus, is_multi_series }
+  return {
+    axis,
+    points,
+    data,
+    focus,
+    // eslint-disable-next-line camelcase
+    is_multi_series,
+    // eslint-disable-next-line camelcase
+    x_value_type: rawData.x_value_type || '',
+    isTimeSeries,
+  }
 })
 
 const containerNode = ref<HTMLElement>()
@@ -179,17 +221,24 @@ const brushExtent = ref<d3.BrushSelection>()
 const limit = ref(DEFAULT_LIMIT)
 const focus = ref<Focus>()
 const shouldAnimate = ref(false)
-const xDomain = ref([0, 1])
-const yDomain = ref([0, 1])
+const xDomain = ref<number[] | Date[]>([0, 1])
+const yDomain = ref<number[] | Date[]>([0, 1])
 const selectionEnabled = ref(false)
 
 const isBrushing = computed(() => brushExtent.value != null)
+
+function axisD3Scale(axis: AxisConfiguration | undefined) {
+  return axis?.scale === ScaleType.Logarithmic ? d3.scaleLog() : d3.scaleLinear()
+}
+
 const xScale = computed(() =>
   axisD3Scale(data.value.axis.x).domain(xDomain.value).range([0, boxWidth.value]),
 )
 const yScale = computed(() =>
   axisD3Scale(data.value.axis.y).domain(yDomain.value).range([boxHeight.value, 0]),
 )
+
+const xScaleTime = computed(() => d3.scaleTime().domain(xDomain.value).range([0, boxWidth.value]))
 
 const symbol: d3.Symbol<unknown, Point> = d3.symbol()
 
@@ -212,7 +261,17 @@ const height = computed(() => config.size.y)
 
 const boxWidth = computed(() => Math.max(0, width.value - margin.value.left - margin.value.right))
 const boxHeight = computed(() => Math.max(0, height.value - margin.value.top - margin.value.bottom))
-const xTicks = computed(() => boxWidth.value / 40)
+const xTicks = computed(() => {
+  switch (data.value.x_value_type) {
+    case 'Time':
+    case 'Date':
+    case 'Date_Time':
+      return boxWidth.value / 80
+    default:
+      return boxWidth.value / 40
+  }
+})
+
 const yTicks = computed(() => boxHeight.value / 20)
 const xLabelLeft = computed(
   () =>
@@ -228,6 +287,16 @@ const yLabelLeft = computed(
 )
 const yLabelTop = computed(() => -margin.value.left + 15)
 const showYLabelText = computed(() => !data.value.is_multi_series)
+const xTickFormat = computed(() => {
+  switch (data.value.x_value_type) {
+    case 'Time':
+      return '%H:%M:%S'
+    case 'Date':
+      return '%d/%m/%Y'
+    default:
+      return '%d/%m/%Y %H:%M:%S'
+  }
+})
 
 watchEffect(() => {
   const boundsExpression =
@@ -258,15 +327,14 @@ const extremesAndDeltas = computed(() => {
     const allYValues = series.flatMap((s) =>
       data.value.data.map((d) => d[s as keyof Point]),
     ) as number[]
-
     yMin = Math.min(...allYValues)
     yMax = Math.max(...allYValues)
   } else {
     ;[yMin = 0, yMax = 0] = d3.extent(data.value.data, (point) => point.y)
   }
 
-  const [xMin = 0, xMax = 0] = d3.extent(data.value.data, (point) => point.x)
-  const dx = xMax - xMin
+  const [xMin = 0, xMax = 0] = d3.extent(data.value.data, (point) => point.x as number)
+  const dx = Number(xMax) - Number(xMin)
   const dy = yMax - yMin
   const paddingX = 0.1 * dx
   const paddingY = 0.1 * dy
@@ -427,16 +495,6 @@ function matchShape(d: Point) {
   return d.shape != null ? SHAPE_TO_SYMBOL[d.shape] ?? d3.symbolCircle : d3.symbolCircle
 }
 
-/** Construct either a linear or a logarithmic D3 scale.
- *
- * The scale kind is selected depending on update contents.
- *
- * @param axis Axis information as received in the visualization update.
- * @returns D3 scale. */
-function axisD3Scale(axis: AxisConfiguration | undefined) {
-  return axis?.scale != null ? SCALE_TO_D3_SCALE[axis.scale]() : d3.scaleLinear()
-}
-
 watchEffect(() => {
   // Update the axes in d3.
   const { xMin, xMax, yMin, yMax, paddingX, paddingY, dx, dy } = extremesAndDeltas.value
@@ -447,7 +505,7 @@ watchEffect(() => {
     xDomain.value = [focus_.x - newPaddingX, focus_.x + newPaddingX]
     yDomain.value = [focus_.y - newPaddingY, focus_.y + newPaddingY]
   } else {
-    xDomain.value = [xMin - paddingX, xMax + paddingX]
+    xDomain.value = [Number(xMin) - paddingX, Number(xMax) + paddingX]
     yDomain.value = [yMin - paddingY, yMax + paddingY]
   }
 })
@@ -458,12 +516,16 @@ watchEffect(() => {
 
 // === Update x axis ===
 
-watchPostEffect(() =>
-  d3XAxis.value
-    .transition()
-    .duration(animationDuration.value)
-    .call(d3.axisBottom(xScale.value).ticks(xTicks.value)),
-)
+watchPostEffect(() => {
+  const xCallVal =
+    data.value.isTimeSeries ?
+      d3
+        .axisBottom<Date>(xScaleTime.value)
+        .ticks(xTicks.value)
+        .tickFormat(d3.timeFormat(xTickFormat.value))
+    : d3.axisBottom(xScale.value).ticks(xTicks.value)
+  return d3XAxis.value.transition().duration(animationDuration.value).call(xCallVal)
+})
 
 // === Update y axis ===
 
@@ -480,6 +542,7 @@ function getPlotData(data: Data) {
     const series = Object.keys(axis).filter((s) => s != 'x')
     const transformedData = series.flatMap((s) =>
       data.data.map((d) => ({
+        ...d,
         x: d.x,
         y: d[s as keyof Point],
         series: s,
@@ -490,6 +553,20 @@ function getPlotData(data: Data) {
   return data.data
 }
 
+function formatXPoint(x: Date | number | DateObj) {
+  if (data.value.isTimeSeries && x instanceof Date) {
+    switch (data.value.x_value_type) {
+      case 'Time':
+        return x.toTimeString()
+      case 'Date':
+        return x.toDateString()
+      default:
+        return x.toString()
+    }
+  }
+  return x
+}
+
 function getTooltipMessage(point: Point) {
   if (data.value.is_multi_series) {
     const axis = data.value.axis
@@ -497,15 +574,15 @@ function getTooltipMessage(point: Point) {
       point.series && point.series in axis ?
         axis[point.series as keyof AxesConfiguration].label
       : ''
-    return `${point.x}, ${point.y}, ${label}`
+    return `${formatXPoint(point.x)}, ${point.y}, ${label}`
   }
-  return `${point.x}, ${point.y}`
+  return `${formatXPoint(point.x)}, ${point.y}`
 }
 
 // === Update contents ===
 
 watchPostEffect(() => {
-  const xScale_ = xScale.value
+  const xScale_ = data.value.isTimeSeries ? xScaleTime.value : xScale.value
   const yScale_ = yScale.value
   const plotData = getPlotData(data.value) as Point[]
   const series = Object.keys(data.value.axis).filter((s) => s != 'x')
@@ -530,7 +607,7 @@ watchPostEffect(() => {
       symbol.type(matchShape).size((d) => (d.size ?? 0.15) * SIZE_SCALE_MULTIPLER),
     )
     .style('fill', (d) => colorScale(d.series || ''))
-    .attr('transform', (d) => `translate(${xScale_(d.x)}, ${yScale_(d.y)})`)
+    .attr('transform', (d) => `translate(${xScale_(Number(d.x))}, ${yScale_(d.y)})`)
   if (data.value.points.labels === VISIBLE_POINTS) {
     d3Points.value
       .selectAll<SVGPathElement, unknown>('text')
@@ -539,7 +616,7 @@ watchPostEffect(() => {
       .transition()
       .duration(animationDuration.value)
       .text((d) => d.label ?? '')
-      .attr('x', (d) => xScale_(d.x) + POINT_LABEL_PADDING_X_PX)
+      .attr('x', (d) => xScale_(Number(d.x)) + POINT_LABEL_PADDING_X_PX)
       .attr('y', (d) => yScale_(d.y) + POINT_LABEL_PADDING_Y_PX)
   }
 })
@@ -619,8 +696,8 @@ function zoomToSelected(override?: boolean) {
     bounds.value = undefined
     limit.value = DEFAULT_LIMIT
     xDomain.value = [
-      extremesAndDeltas.value.xMin - extremesAndDeltas.value.paddingX,
-      extremesAndDeltas.value.xMax + extremesAndDeltas.value.paddingX,
+      Number(extremesAndDeltas.value.xMin) - extremesAndDeltas.value.paddingX,
+      Number(extremesAndDeltas.value.xMax) + extremesAndDeltas.value.paddingX,
     ]
     yDomain.value = [
       extremesAndDeltas.value.yMin - extremesAndDeltas.value.paddingY,