+++ app/gui2/src/components/visualizations/TableVisualization.vue
@@ -8,10 +8,10 @@ import { useVisualizationConfig } from '@/util/visualizationBuiltins'
 import type {
   CellClassParams,
   CellClickedEvent,
+  ColDef,
   ICellRendererParams,
   SortChangedEvent,
-} from 'ag-grid-community'
-import type { ColDef } from 'ag-grid-enterprise'
+} from 'ag-grid-enterprise'
 import { computed, onMounted, ref, shallowRef, watchEffect, type Ref } from 'vue'
 
 export const name = 'Table'